feat: integrate autocomplete
