+++ src/routes/__root.tsx
@@ -5,7 +5,6 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { createRootRoute, Outlet } from "@tanstack/react-router";
 
 import { ManaqlClient } from "@/api/manaql/client";
-import { ScryfallClient } from "@/api/scryfall/client";
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
 import { Toaster } from "@/components/ui/toaster";
@@ -18,7 +17,6 @@ const apolloClient = new ApolloClient({
 });
 
 export const manaqlClient = new ManaqlClient(apolloClient);
-export const scryfallClient = new ScryfallClient();
 export const queryClient = new QueryClient();
 
 export const Route = createRootRoute({
